all stories as stateless func
