+++ src/components/Toggle/index.js
@@ -1,1 +1,50 @@
-export { default } from './Toggle';
+/**
+ *
+ * Toggle
+ *
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import {
+  commonDefaultProps,
+  commonPropTypes,
+} from '../../commonPropTypes/input';
+
+import Label from '../Label';
+
+import StyledToggle, { ToggleWrapper } from '../../styled/Toggle';
+
+function Toggle({ id, name, onChange, value }) {
+  return (
+    <ToggleWrapper>
+      <Label htmlFor={id || name}>
+        <StyledToggle
+          value={value}
+          id={id || name}
+          name={id || name}
+          onChange={e => {
+            onChange({ target: { name, value: e.target.checked } });
+          }}
+        />
+        <span>OFF</span>
+        <span>ON</span>
+      </Label>
+    </ToggleWrapper>
+  );
+}
+
+Toggle.defaultProps = {
+  ...commonDefaultProps,
+  onChange: () => {},
+  value: false,
+};
+
+Toggle.propTypes = {
+  ...commonPropTypes,
+  name: PropTypes.string.isRequired,
+  onChange: PropTypes.func,
+  value: PropTypes.bool,
+};
+
+export default Toggle;
