+++ src/index.js
@@ -1,4 +1,3 @@
 export { default as Button } from './components/Button';
 export { default as Textarea } from './components/Textarea';
 export { default as Label } from './components/Label';
-export { default as GlobalStyles } from './components/GlobalStyles';
