+++ src/styled/Enumeration/index.js
@@ -1,0 +1,8 @@
+/**
+ *
+ * Enumeration
+ *
+ */
+
+export { default } from './Enumeration';
+export { default as EnumerationWrapper } from './EnumerationWrapper';
