+++ src/styled/Toggle/index.js
@@ -1,0 +1,8 @@
+/**
+ *
+ * Toggle
+ *
+ */
+
+export { default } from './Toggle';
+export { default as ToggleWrapper } from './ToggleWrapper';
