+++ stories/Compos/InputSearch.stories.js
@@ -1,38 +1,24 @@
-import React from 'react';
+import React, { useState } from 'react';
 /* eslint-disable import/no-extraneous-dependencies */
 import { storiesOf } from '@storybook/react';
 import InputText from '../../src/components/InputText';
 
-const defaultProps = {
-  placeholder: 'Tap something to search…',
-  name: 'input',
-  type: 'search',
-};
+function InputTextStory() {
+  const [val, setValue] = useState('');
 
-class InputTextStory extends React.Component {
-  state = {
-    value: {
-      input: '',
-    },
+  const defaultProps = {
+    placeholder: 'Tap something to search…',
+    name: 'input',
+    type: 'search',
   };
 
-  handleChange = ({ target }) => {
-    this.setState(prevState => ({
-      value: { ...prevState.value, [target.name]: target.value },
-    }));
-  };
-
-  render() {
-    const { value } = this.state;
-
-    return (
-      <InputText
-        {...defaultProps}
-        onChange={this.handleChange}
-        value={value.input}
-      />
-    );
-  }
+  return (
+    <InputText
+      {...defaultProps}
+      onChange={({ target: { value } }) => setValue(value)}
+      value={val}
+    />
+  );
 }
 
 storiesOf('Components|InputSearch', module).add('Default', () => (
